move performance tests into separate file also add some even more descriptive comments
